refactor: Add getEvents function to comManager.js and update events page
This commit refactors the comManager.js file by adding a new function called getEvents. This function fetches events data from the server and updates the events state in the counter store. Additionally, the events page (index.vue) is updated to use the computed property to retrieve the events from the store. This improves code organization and separation of concerns.