feat: Add functions to fetch followers and followed users
The code changes include adding two new functions, `getFollowers()` and `getFollowed()`, to the `userManager` module. These functions make API calls to fetch the count of followers and followed users for a specific user. The functions use the `axios` library and require the user's ID and token for authentication.

This commit message suggests adding the new functions to the `userManager` module to fetch followers and followed users.