feat: Add logout function to user manager
The code changes include adding a new function, `logout()`, to the `userManager` module. This function makes an API call to log out the user by sending a `POST` request to the `/logout` endpoint. The function also clears the user's information and sets the `loggedIn` flag to `false`.

This commit message suggests adding the `logout()` function to the `userManager` module to handle user logout functionality.