refactor: Update login page with registration link
This commit updates the login page by adding a registration link. The link is styled with a gradient background and text that transitions from orange to yellow. Users can click on the link to navigate to the registration page. This enhancement improves the user experience by providing an easy way for new users to create an account.